move context to utils
